+++ src/app.ts
@@ -1,7 +1,7 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
+import { AppService } from "@application/services/app.service";
+import { OptsService } from "@application/services/opts.service";
 import Container, { Service } from "typedi";
-import { AppService } from "./services/app.service";
-import { OptsService } from "./services/opts.service";
 
 @Service()
 export class App {
