+++ src/application/services/app.service.ts
@@ -1,11 +1,11 @@
+import { IExecutable } from "@domain/interfaces/executable.interface";
+import { Ui } from "@interface/ui";
 import { Container, Service } from "typedi";
-import { IExecutable } from "../contracts/executable.interface";
-import { AnyType } from "../types";
-import { Ui } from "../ui";
+import { AnyType } from "../../types";
 import { filePathEnvReplace } from "../utils/file-path-env-replace";
 import { progressBarExecutableMetadata } from "../utils/progress-bar-executable-metadata";
-import { ExternalFileService } from "./external-file.service";
 import { ExecutableService } from "./executable-service";
+import { ExternalFileService } from "./external-file.service";
 import { ProcessesScanner, ScheduledTasksScanner, ServicesScanner } from "./scanners";
 import { IScanner } from "./scanners/scanner.interface";
 import { TOptions } from "./types";
