+++ src/application/services/external-file.service.ts
@@ -1,9 +1,9 @@
+import { IExecutable } from "@domain/interfaces/executable.interface";
 import { mkdir, readFile, writeFile } from "fs/promises";
 import { join } from "path";
 import { Service } from "typedi";
-import { IExecutable } from "../contracts/executable.interface";
-import { HashStatus, OutputJson } from "./types";
 import { isNullOrUndefined } from "../utils/is-null-or-undefined";
+import { HashStatus, OutputJson } from "./types";
 
 @Service()
 export class ExternalFileService {
