+++ src/application/services/scanners/__test__/scheduled-tasks.scrapper.spec.ts
