+++ src/application/services/scanners/processes.scanner.ts
@@ -1,10 +1,10 @@
+import { Executable } from "@domain/entities/executable.entity";
+import { ExecutableSource } from "@domain/enum/executable-source.enum";
+import { IExecutable } from "@domain/interfaces/executable.interface";
 import { basename } from "node:path";
 import { Inject, Service } from "typedi";
-import { IExecutable } from "../../contracts/executable.interface";
-import { Executable } from "../../entities/executable.entity";
 import { Powershell } from "../../utils/powershell";
 import { IScanner } from "./scanner.interface";
-import { ExecutableSource } from "../../enum/executable-source.enum";
 
 @Service()
 export class ProcessesScanner implements IScanner {
@@ -30,7 +30,7 @@ export class ProcessesScanner implements IScanner {
 
       const processBaseName = basename(filePath);
 
-      executables.add(new Executable(processBaseName, filePath, ExecutableSource.PROCESS));
+      executables.add(new Executable(processBaseName, filePath, ExecutableSource.PROCESS, ""));
     }
 
     return Array.from(executables);
