+++ src/application/services/scanners/scanner.interface.ts
@@ -1,5 +1,5 @@
-import { IExecutable } from "../../contracts/executable.interface";
-import { TExecutableSource } from "../../enum/executable-source.enum";
+import { TExecutableSource } from "@domain/enum/executable-source.enum";
+import { IExecutable } from "@domain/interfaces/executable.interface";
 
 export interface IScanner {
   name: string;
