+++ src/application/services/scanners/scheduled-tasks.scanner.ts
@@ -1,9 +1,9 @@
+import { Executable } from "@domain/entities/executable.entity";
+import { ExecutableSource } from "@domain/enum/executable-source.enum";
+import { IExecutable } from "@domain/interfaces/executable.interface";
 import { Service } from "typedi";
-import { IExecutable } from "../../contracts/executable.interface";
-import { IScanner } from "./scanner.interface";
 import { Powershell } from "../../utils/powershell";
-import { Executable } from "../../entities/executable.entity";
-import { ExecutableSource } from "../../enum/executable-source.enum";
+import { IScanner } from "./scanner.interface";
 
 @Service()
 export class ScheduledTasksScanner implements IScanner {
@@ -31,7 +31,7 @@ export class ScheduledTasksScanner implements IScanner {
         continue;
       }
       const path = match[0].substring(2).replaceAll('"', "");
-      executables.push(new Executable("Scheduled Task", path, ExecutableSource.SCHEDULED_TASK));
+      executables.push(new Executable("Scheduled Task", path, ExecutableSource.SCHEDULED_TASK, ""));
     }
 
     return executables;
