+++ src/application/services/scanners/services.scranner.ts
@@ -1,8 +1,8 @@
+import { Executable } from "@domain/entities/executable.entity";
+import { ServiceExecutable } from "@domain/entities/service-executable.entity";
+import { ExecutableSource } from "@domain/enum/executable-source.enum";
+import { IExecutable } from "@domain/interfaces/executable.interface";
 import { Service } from "typedi";
-import { IExecutable } from "../../contracts/executable.interface";
-import { Executable } from "../../entities/executable.entity";
-import { ServiceExecutable } from "../../entities/service-executable.entity";
-import { ExecutableSource } from "../../enum/executable-source.enum";
 import { Powershell } from "../../utils/powershell";
 import { IScanner } from "./scanner.interface";
