+++ src/application/services/types.ts
@@ -1,4 +1,4 @@
-import { IExecutable } from "../contracts/executable.interface";
+import { IExecutable } from "@domain/interfaces/executable.interface";
 
 export type TOptions = {
   inputFile?: string;
