+++ src/application/utils/__test__/powershell.spec.ts
