+++ src/application/utils/file-path-env-replace.ts
