+++ src/application/utils/is-null-or-undefined.ts
