+++ src/application/utils/progress-bar-executable-metadata.ts
@@ -1,4 +1,4 @@
-import { IExecutable } from "../contracts/executable.interface";
+import { IExecutable } from "@domain/interfaces/executable.interface";
 
 export function progressBarExecutableMetadata(allExecutables: IExecutable[]): Record<string, number> {
   const metadata: Record<string, number> = {};
