+++ src/domain/entities/executable.entity.ts
@@ -1,5 +1,5 @@
-import { IExecutable } from "../contracts/executable.interface";
-import { ExecutableSource, TExecutableSource } from "../enum/executable-source.enum";
+import { ExecutableSource, TExecutableSource } from "@domain/enum/executable-source.enum";
+import { IExecutable } from "@domain/interfaces/executable.interface";
 
 export class Executable implements IExecutable {
   public name: string;
