+++ src/domain/entities/service-executable.entity.ts
@@ -1,6 +1,6 @@
+import { getPathAndArgs } from "@application/utils/get-path-and-args";
+import { NonFunctionProperties } from "../../types";
 import { ExecutableSource } from "../enum/executable-source.enum";
-import { NonFunctionProperties } from "../types";
-import { getPathAndArgs } from "../utils/get-path-and-args";
 import { Executable } from "./executable.entity";
 
 export class ServiceExecutable extends Executable {
