+++ src/domain/enum/executable-source.enum.ts
@@ -1,4 +1,4 @@
-import { ObjectValues } from "../types";
+import { ObjectValues } from "../../types";
 
 export const ExecutableSource = {
   SERVICE: "Service",
