+++ src/domain/interfaces/executable.interface.ts
